TFM/Deployment/UI/requirements.txt
streamlit
requests
pandas
matplotlib
pandas_gbq
fastapi
uvicorn
google-cloud-bigquery
joblib
db-dtypes